+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/background/BackgroundJobParameters.kt
@@ -1,4 +1,4 @@
-package com.outsystems.plugins.healthfitnesslib.background
+package com.outsystems.plugins.healthfitness.background
 
 import com.google.gson.annotations.SerializedName
 
@@ -9,6 +9,8 @@ data class BackgroundJobParameters (
     @SerializedName("TimeUnit") val timeUnit : String? = null,
     @SerializedName("TimeUnitGrouping") val timeUnitGrouping : Int? = 1,
     @SerializedName("JobFrequency") val jobFrequency : String? = null,
+    @SerializedName("NotificationFrequency") val notificationFrequency : String? = null,
+    @SerializedName("NotificationFrequencyGrouping") val notificationFrequencyGrouping : Int? = 1,
     @SerializedName("NotificationHeader") val notificationHeader : String,
     @SerializedName("NotificationBody") val notificationBody : String
 )