+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/background/database/BackgroundJobDao.kt
@@ -1,26 +1,29 @@
-package com.outsystems.plugins.healthfitnesslib.background.database
+package com.outsystems.plugins.healthfitness.background.database
 
-import androidx.room.Dao
-import androidx.room.Delete
-import androidx.room.Insert
-import androidx.room.Query
+import androidx.room.*
 
 @Dao
 interface BackgroundJobDao {
 
-    @Query("SELECT * FROM backgroundJob")
+    @Query("SELECT * FROM ${BackgroundJob.TABLE_NAME}")
     fun getAll(): List<BackgroundJob>
 
-    //@Query("SELECT * FROM backgroundJob WHERE variable = :variable AND comparison = :comparison AND value = :value")
-    //fun findByPrimaryKey(variable: String, comparison: String, value: Float)
+    @Query("SELECT COUNT(*) FROM ${BackgroundJob.TABLE_NAME} WHERE variable = :name")
+    fun getBackgroundJobCountForVariable(name : String): Int
 
-    @Query("SELECT * FROM backgroundJob WHERE variable = :name")
+    @Query("SELECT * FROM ${BackgroundJob.TABLE_NAME} WHERE id = :id")
+    fun findById(id: String): BackgroundJob?
+
+    @Query("SELECT * FROM ${BackgroundJob.TABLE_NAME} WHERE variable = :name")
     fun findByVariableName(name : String): List<BackgroundJob>
 
     @Insert
-    fun insert(backgroundJob: BackgroundJob) : Long
+    fun insert(backgroundJob: BackgroundJob): Long
 
     @Delete
     fun delete(backgroundJob: BackgroundJob)
+
+    @Update
+    fun update(backgroundJob: BackgroundJob)
 
 }