+++ src/android/com/outsystems/plugins/healthfitness/oshealthfitnesslibrary/store/AdvancedQueryResponse.kt
@@ -1,4 +1,4 @@
-package com.outsystems.plugins.healthfitnesslib.store
+package com.outsystems.plugins.healthfitness.store
 
 import com.outsystems.plugins.healthfitness.HealthFitnessError