+++ www/OSHealthFitness.js
@@ -39,3 +39,15 @@ exports.getLastRecord = function (success, error, variable) {
 exports.setBackgroundJob = function (success, error, params) {
     exec(success, error, 'OSHealthFitness', 'setBackgroundJob', [params]);
 };
+
+exports.deleteBackgroundJob = function (success, error, params) {
+    exec(success, error, 'OSHealthFitness', 'deleteBackgroundJob', [params]);
+};
+
+exports.listBackgroundJobs = function (success, error) {
+    exec(success, error, 'OSHealthFitness', 'listBackgroundJobs');
+};
+
+exports.updateBackgroundJob = function (success, error, params) {
+    exec(success, error, 'OSHealthFitness', 'updateBackgroundJob', [params]);
+};